+++ ThirdPartyAdapters/maio/maio/src/main/java/jp/maio/sdk/android/mediation/admob/adapter/FirstLoadInterface.java
@@ -1,9 +1,0 @@
-package jp.maio.sdk.android.mediation.admob.adapter;
-
-/**
- * Interface to capture the first ad loaded event.
- */
-
-public interface FirstLoadInterface {
-    public void adLoaded(String zoneId);
-}