+++ ThirdPartyAdapters/facebook/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Facebook Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 5.3.0.0
+- Updated mediation service name for Google Mobile Ads.
+- Added adapter version to the initialization call.
+- Verified compatibility with Facebook SDK v5.3.0.
+
+## 5.2.0.2
+- Added support for flexible banner ad sizes.
+
 ## 5.2.0.1
 - Updated adapter to support new open-beta Rewarded API.
 - Updated the minimum required Google Mobile Ads SDK version to 17.2.0.